Working with large matrices

+++ code2/Helpers.cpp
@@ -7,7 +7,7 @@ using namespace std;
 // IMPORTANT! SQUARE MATRICES ONLY! (as per the spec)
 matrix subt(matrix M1, matrix M2) {
     int d = (int) M1.size();
-    matrix result (d, vector<int> (d, 0));
+    matrix result (d, vector<long long> (d, 0));
     
     for(int i = 0; i < d; i++) {
         for(int j = 0; j < d; j++) {
@@ -20,7 +20,7 @@ matrix subt(matrix M1, matrix M2) {
 matrix add(matrix M1, matrix M2) {
     int d = (int) M1.size();
     
-    matrix result (d, vector<int> (d, 0));
+    matrix result (d, vector<long long> (d, 0));
     for(int i = 0; i < d; i++) {
         for(int j = 0; j < d; j++) {
             result[i][j] = M1[i][j] + M2[i][j];
@@ -31,7 +31,7 @@ matrix add(matrix M1, matrix M2) {
 
 matrix mult(matrix M1, matrix M2) {
     int d = (int) M1.size();
-    matrix result (d, vector<int> (d, 0));
+    matrix result (d, vector<long long> (d, 0));
     
     for(int i = 0; i < d; i++) {
         for(int j = 0; j < d; j++) {
